move ErgoScriptCompiler to utils package;
